refactor(actions): move action imports to a centralized lib directory
